backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic
pandas
pyarrow
dask[dataframe]
plotly
redis
celery[redis]
tenacity
numexpr
alembic
pytest
